Separate Routes and handlers from app
