+++ src/tracer/src/daemon/mod.rs
@@ -1,5 +1,7 @@
 pub mod app;
 pub mod client;
 pub mod daemon_run;
+pub mod handlers;
+pub mod routes;
 pub mod server;
 pub mod structs;
